// src/app/api/models/registro-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RegistroRequest {
  contrasena?: string;
  correo?: string;
  institutionId?: number;
  nombreCompleto?: string;
  nombreUsuario?: string;
  rol?: 'MASTER' | 'ADMIN' | 'CAJERO' | 'CLIENTE' | 'SISTEMA' | 'TUTOR' | 'CONSULTOR';
}